+++ src/components/DatePicker/DatePicker.tsx
@@ -2,8 +2,6 @@ import cx from 'classnames';
 import * as dayjs from 'dayjs';
 import * as isoWeek from 'dayjs/plugin/isoWeek';
 import * as localeData from 'dayjs/plugin/localeData';
-import * as timezone from 'dayjs/plugin/timezone';
-import * as utc from 'dayjs/plugin/utc';
 import * as React from 'react';
 
 import { FontAwesomeIcon } from '../FontAwesomeIcon/FontAwesomeIcon';
@@ -13,8 +11,6 @@ import { TimePicker } from './TimePicker';
 
 dayjs.extend(isoWeek);
 dayjs.extend(localeData);
-dayjs.extend(utc);
-dayjs.extend(timezone);
 
 export interface DatePickerProps {
   className?: string;
@@ -24,24 +20,20 @@ export interface DatePickerProps {
   use24Hour?: boolean;
   showTimePicker?: boolean;
   showSeconds?: boolean;
-  timezone?: string;
   onChange: (date: Date) => void;
 }
 
 export const DatePicker = ({
-  date,
+  date = new Date(),
   startDate,
   endDate,
   showTimePicker = true,
   showSeconds = true,
   use24Hour,
   className,
-  timezone,
   onChange
 }: DatePickerProps): React.ReactElement => {
-  const inputDate = dayjs(date) || new Date();
-  const currentTz = timezone || dayjs.tz.guess();
-  const [currentDate, setCurrentDate] = React.useState(dayjs.tz(inputDate, currentTz));
+  const [currentDate, setCurrentDate] = React.useState(dayjs(date));
   const getMonthMatrix = () => {
     const monthStartAtDay = currentDate.date(0).day();
     const daysInMonth = currentDate.daysInMonth();
@@ -71,7 +63,7 @@ export const DatePicker = ({
 
   const selectDate = (dayOfMonth: number) => {
     setCurrentDate(prev => prev.date(dayOfMonth));
-    if (onChange) onChange(currentDate.tz(timezone).date(dayOfMonth).toDate());
+    if (onChange) onChange(currentDate.date(dayOfMonth).toDate());
   };
   const goNextMonth = () => {
     setCurrentDate(prev => prev.add(1, 'month'));
@@ -173,7 +165,7 @@ export const DatePicker = ({
                         className={cx('dui-datepicker-day', {
                           active:
                             selectedDayOfMonth === day &&
-                            dayjs(inputDate).month() === currentDate.month()
+                            dayjs(date).month() === currentDate.month()
                         })}
                         onClick={() => selectDate(parseInt(day.toString()))}
                       >
@@ -189,7 +181,6 @@ export const DatePicker = ({
       </table>
       {showTimePicker && (
         <TimePicker
-          timezone={timezone}
           showSeconds={showSeconds}
           use24Hour={use24Hour}
           date={currentDate}